requirements.txt
pyyaml
pytz
eventbrite